+++ src/core/init/interface.ts
@@ -7,13 +7,10 @@
  */
 
 import type { InitialRoute } from 'core/router';
+import type { CookieStore } from 'core/cookies';
 
 import type { State } from 'core/component/state';
 import type { ComponentOptions } from 'core/component/engines';
-
-type OptionalState = {
-	[K in keyof State]?: State[K];
-};
 
 export interface AppSSR {
 	content: string;
@@ -22,17 +19,25 @@ export interface AppSSR {
 
 export type App = Element | AppSSR;
 
-export interface InitAppOptions extends OptionalState {
+export interface CreateAppOptions {
 	/**
-	 * The unique application identifier
+	 * A function that is called before the initialization of the root component
+	 * @param rootComponentParams
 	 */
-	appId?: string;
+	setup?(rootComponentParams: ComponentOptions): void;
 
 	/**
 	 * A link to the element where the application should be mounted.
 	 * This parameter is only used when initializing the application in a browser.
 	 */
 	targetToMount?: Nullable<HTMLElement>;
+}
+
+export type InitAppOptions = CreateAppOptions & Overwrite<State, {
+	/**
+	 * A store of application cookies
+	 */
+	cookies: CookieStore;
 
 	/**
 	 * The initial route for initializing the router.
@@ -41,19 +46,7 @@ export interface InitAppOptions extends OptionalState {
 	route?: InitialRoute;
 
 	/**
-	 * A function that is called before the initialization of the root component
-	 * @param rootComponentParams
+	 * An API to work with a network, such as testing of the network connection, etc.
 	 */
-	setup?(rootComponentParams: ComponentOptions): void;
-
-	/**
-	 * Sets the passed flag to a ready status.
-	 * When all the declared flags are ready, the application itself will be initialized.
-	 *
-	 * @param flag
-	 */
-	ready(flag: string): Promise<(
-		rootComponentName: Nullable<string>,
-		opts: InitAppOptions
-	) => Promise<App>>;
-}
+	net?: State['net'];
+}>;